add if command

+++ src/Command/PackageExpiredCommand.php
@@ -48,11 +48,12 @@ class PackageExpiredCommand extends Command
 
         foreach ($allPackages as $package) {
             $packageDateDelete = $package->getExpireOn();
-
-            if ($packageDateDelete < $currentDate) {
-                $this->packageRepository->remove($package);
-                $text = 'Le package ' . $package->getName() . ' a bien été supprimé';
-                $io->success($text);
+            if ($packageDateDelete !== null) {
+                if ($packageDateDelete < $currentDate) {
+                    $this->packageRepository->remove($package);
+                    $text = 'Le package ' . $package->getName() . ' a bien été supprimé';
+                    $io->success($text);
+                }
             }
         }
